move Inprogress file to common
