+++ Starteryou/frontend/src/pages/AboutPage.jsx
@@ -5,7 +5,7 @@ import OurVision from "../components/About/OurVision";
 import Team from "../components/About/Team";
 import TechTeam from "../components/About/TechTeam";
 import Contact from "../components/Landing/Contact";
-import InProgressPage from "./InProgressPage";
+import InProgressPage from "../components/Common/InProgressPage";
 export default function AboutPage() {
   const deployReady = false;
   return deployReady ? (
