+++ Starteryou/frontend/src/pages/EducationPage.jsx
@@ -5,7 +5,7 @@ import StartJourney from "../components/Education/StartJourney";
 import Testimonials from "../components/Education/Testimonials";
 import UnlockPotential from "../components/Education/UnlockPotential";
 import YourJourney from "../components/Education/YourJourney";
-import InProgressPage from "./InProgressPage";
+import InProgressPage from "../components/Common/InProgressPage";
 export default function EducationPage() {
   const deployReady = false;
   return deployReady ? (
