+++ Starteryou/frontend/src/pages/JobPortalPage.jsx
@@ -6,7 +6,7 @@ import LatestInsight from "../components/JobPortal/LatestInsight";
 import LaunchBanner from "../components/JobPortal/LaunchBanner";
 import PathStart from "../components/JobPortal/PathStart";
 import UnlockPotential from "../components/JobPortal/UnlockPotential";
-import InProgressPage from "./InProgressPage";
+import InProgressPage from "../components/Common/InProgressPage";
 export default function JobPortalPage() {
   const deployReady = true;
   return deployReady ? (
